correcion delete complaint

+++ src/components/ButtonActionComplaint/index.jsx
@@ -4,16 +4,19 @@ import IconButton from '@material-ui/core/IconButton';
 import {HowToReg, Delete} from '@material-ui/icons';
 import complaintHttpClient from '../../services/Api/complaint.httpClient';
 import {useComplaintsUpdate} from '../../hooks/useComplaintsUpdate';
-import { newVoteSessionStorage, isExistVote } from '../../helpers/sessionStorage';
+import { newVoteSessionStorage, isExistVote, isDeleted } from '../../helpers/sessionStorage';
 
-const ButtonActionComplaint = ({userC, id, setVote}) => {
+const ButtonActionComplaint = ({userC, id, setVote, state}) => {
     const user = useUser();
     const { setClick } = useComplaintsUpdate();
-    const [color, setcolor] = useState(!isExistVote(id) ? 'secondary' : 'inherit')
+    const [color, setcolor] = useState(!isExistVote(id) ? 'secondary' : 'inherit');
+    const colorDelete = isDeleted(state) ? 'secondary' : 'inherit';
 
     const handleDelete = () => {
-        complaintHttpClient.deleteId(id);
-        setClick(id);
+        if(isDeleted(state)){
+            complaintHttpClient.deleteId(id);
+            setClick(id);
+        }
     }
 
     const handleVote = () => {
@@ -28,7 +31,7 @@ const ButtonActionComplaint = ({userC, id, setVote}) => {
     const buttonAction = () => {
         if(userC !== null){
             return userC.ID_U === user.uid  ? <IconButton aria-label="add to favorites" onClick={handleDelete}>
-                                                <Delete color='secondary'/>
+                                                <Delete color={colorDelete}/>
                                               </IconButton> 
                                             : <IconButton aria-label="add to favorites" onClick={handleVote} >
                                                 <HowToReg color={color}/>
